Move methods from the logger to the thread safe lib
These methods were added on 629efb605728b31ad9644f6f0acaf3760b641a29 to
the logger silence, and then moved to the logger on
2518bda97cbbcb33dc9a92e70d5b01c09e64d12d. They only exist because we
need to call the `#level` method, which handles the thread safety,
instead of accessing the underlying `@level` variable, so lets keep them
within the module that actually adds the thread safety code instead.

This also changes the thread safety module to be prepended instead, so
we can simply implement `initialize` and call `super` from the module
itself, instead of having to handle another `after_initialize` callback.

We can also remove the inclusion of the concern module, as it does not
add any value to that internal thread safety module.